Ignore players that have no data in them to be read (#105)
Players which have no data read to be read result in audio being blocked. This can result in choppy
audio if a player is used like so:

    p := context.NewPlayer()
    defer p.Close()
    p.Write(coolSounds)
    doSomethingExpensive()

This is counter-intuitive, as I wouldn't expect the time I call a Close() method to affect user experience in
any way.

This fixes that by not using io.Pipe. Since io.Pipe synchronizes reads and writes, it is not an ideal choice
for data transfer here, as it is impossible to determine if there is a writer on the other side of a pipe.
Instead, a buffer + mutex to allow concurrent access will suffice.