player: optimize memory re-use by using one buffer pool for all players.
This way different player instances can re-use memory buffers.
This helps keeping the allocation rate low and reduces gc.

Sharing is fine under the assumption that after a short while all buffers are
at least as large as we currently need them to be. If we get a buffer from the pool
that is smaller than the current required buffer size, we'll alloc a bigger buffer
and return that one into the pool. Next time, we'll probably get a buffer of the
correct size.

We now also share the p.buf buffer between players. If a player is stopped or has reached eof,
the buffer is put back into the pool.

In total, this change helps to reduce the allocation rate (and in turn garbage pressure) when
playing a lot of short lived sound effects, and it does not hinder long running playback.